+++ src/azure-pipelines-vscode-ext/index.js
@@ -92,17 +92,25 @@ function activate(context) {
 							var result = null;
 							while(true) {
 								try {
-									result = handle.askForInput ? await vscode.window.showInputBox({
-										ignoreFocusOut: true,
-										placeHolder: "value",
-										value: result,
-										prompt: `${repositoryAndRef} (${filename}) (Leave empty to show a folder picker)`,
-										title: "Provide the uri or filepath to the folder of your required Repository"
-									}) : null;
-									if(result === "") {
-										var uris = await vscode.window.showOpenDialog({ canSelectFolders: true, openLabel: "Select Repository"});
-										result = uris && uris[0] ? uris[0].toString() : null;
-									}
+									result = handle.askForInput ? await new Promise((resolve, reject) => {
+										let inputbox = vscode.window.createInputBox();
+										inputbox.ignoreFocusOut = true;
+										inputbox.placeholder = "local folder path or vscode uri";
+										inputbox.buttons = [{
+											iconPath: new vscode.ThemeIcon("folder"),
+											tooltip: "Show folder picker"
+										}];
+										inputbox.prompt = `${repositoryAndRef} (${filename})`;
+										inputbox.value = result;
+										inputbox.title = "Provide the vscode uri or filepath to the folder of your required Repository";
+										inputbox.onDidAccept(() => resolve(inputbox.value));
+										inputbox.onDidHide(() => resolve(null));
+										inputbox.onDidTriggerButton(async (button) => {
+											var uris = await vscode.window.showOpenDialog({ canSelectFolders: true, openLabel: "Select Repository"});
+											resolve(uris && uris[0] ? uris[0].toString() : null);
+										});
+										inputbox.show();
+									}): null;
 									if(result) {
 										handle.repositories ??= {};
 										handle.repositories[repositoryAndRef] = result;